+++ httpclient-vertx/src/main/java/io/fabric8/kubernetes/client/vertx/VertxHttpRequest.java
@@ -5,18 +5,14 @@ import io.fabric8.kubernetes.client.http.HttpRequest;
 import io.fabric8.kubernetes.client.http.HttpResponse;
 import io.fabric8.kubernetes.client.http.StandardHttpRequest;
 import io.vertx.core.Future;
-import io.vertx.core.Handler;
 import io.vertx.core.MultiMap;
 import io.vertx.core.Vertx;
 import io.vertx.core.buffer.Buffer;
-import io.vertx.core.buffer.impl.VertxByteBufAllocator;
 import io.vertx.core.http.HttpClient;
 import io.vertx.core.http.HttpClientResponse;
 import io.vertx.core.http.RequestOptions;
 import io.vertx.core.streams.ReadStream;
-import io.vertx.core.streams.impl.InboundBuffer;
 
-import java.io.IOException;
 import java.io.InputStream;
 import java.net.URI;
 import java.nio.ByteBuffer;
@@ -30,7 +26,7 @@ import java.util.function.Function;
 
 class VertxHttpRequest implements HttpRequest {
 
-  private final Vertx vertx;
+  final Vertx vertx;
   private final RequestOptions options;
   private final StandardHttpRequest.BodyContent body;
 
@@ -140,102 +136,7 @@ class VertxHttpRequest implements HttpRequest {
           StandardHttpRequest.InputStreamBodyContent bodyContent = (StandardHttpRequest.InputStreamBodyContent) body;
           InputStream is = bodyContent.getContent();
 
-          ReadStream<Buffer> stream = new ReadStream<>() {
-
-            private InboundBuffer<Buffer> inboundBuffer;
-            private Handler<Throwable> exceptionHandler;
-            private Handler<Void> endHandler;
-            byte[] bytes;
-
-            @Override
-            public ReadStream<Buffer> exceptionHandler(Handler<Throwable> handler) {
-              exceptionHandler = handler;
-              return this;
-            }
-            @Override
-            public ReadStream<Buffer> handler(Handler<Buffer> handler) {
-              boolean start = inboundBuffer == null && handler != null;
-              if (start) {
-                inboundBuffer = new InboundBuffer<>(vertx.getOrCreateContext());
-                inboundBuffer.drainHandler(v -> {
-                  readChunk();
-                });
-              }
-              if (handler != null) {
-                inboundBuffer.handler(buff -> {
-                  if (buff == null) {
-                    if (endHandler != null) {
-                      endHandler.handle(null);
-                    }
-                  } else {
-                    handler.handle(buff);
-                  }
-                });
-              } else {
-                inboundBuffer.handler(null);
-              }
-              if (start) {
-                readChunk();
-              }
-              return this;
-            }
-            private void readChunk() {
-              Future<Buffer> fut = vertx.executeBlocking(p -> {
-                if (bytes == null) {
-                  bytes = new byte[1024];
-                }
-                int amount;
-                try {
-                  amount = is.read(bytes);
-                } catch (IOException e) {
-                  p.fail(e);
-                  return;
-                }
-                if (amount == -1) {
-                  p.complete();
-                } else {
-                  p.complete(Buffer.buffer(VertxByteBufAllocator.DEFAULT.heapBuffer(amount, Integer.MAX_VALUE).writeBytes(bytes, 0, amount)));
-                }
-              });
-              fut.onComplete(ar -> {
-                if (ar.succeeded()) {
-                  Buffer chunk = ar.result();
-                  if (chunk != null) {
-                    boolean writable = inboundBuffer.write(chunk);
-                    if (writable) {
-                      readChunk();
-                    } else {
-                      // Full
-                    }
-                  } else {
-                    inboundBuffer.write((Buffer) null);
-                  }
-                } else {
-                  request.reset(0, ar.cause());
-                }
-              });
-            }
-            @Override
-            public ReadStream<Buffer> endHandler(Handler<Void> handler) {
-              endHandler = handler;
-              return this;
-            }
-            @Override
-            public ReadStream<Buffer> pause() {
-              inboundBuffer.pause();
-              return this;
-            }
-            @Override
-            public ReadStream<Buffer> resume() {
-              inboundBuffer.resume();
-              return this;
-            }
-            @Override
-            public ReadStream<Buffer> fetch(long amount) {
-              inboundBuffer.fetch(amount);
-              return this;
-            }
-          };
+          ReadStream<Buffer> stream = new InputStreamReadStream(this, is, request);
           fut = request.send(stream);
         } else {
           fut = Future.failedFuture("Unsupported body content");
@@ -258,4 +159,5 @@ class VertxHttpRequest implements HttpRequest {
     multiMap.names().stream().forEach(k -> headers.put(k, multiMap.getAll(k)));
     return headers;
   }
+
 }
